refactor: separate pages into its own file
