+++ src/App.jsx
@@ -8,7 +8,7 @@ import DatabaseTable from "./pages/DatabaseTable";
 import GoogleAuthSuccess from "./pages/GoogleAuthSuccess";
 import { parseCSVFile, Identifiers } from './utils/CSVParser';
 import Localization from './pages/Localization';
-import ContactDesignation from './pages/ContactDesignation';
+import ContactDesignation from './pages/ContactDesignation/ContactDesignation';
 
 const Tab = ({ title, isActive, onClick, onClose }) => {
     return (
